backend/requirements.txt
fastapi
uvicorn[standard]
chromadb
openai
python-dotenv
numpy

